Colocate selection tests, fix lint warning
- Move selection functional tests to plugins/selection/useXDSTableSelection.test.tsx
- Move selection perf tests to plugins/selection/useXDSTableSelection-perf.test.tsx
- Remove selection tests from XDSTable.test.tsx (65 tests remain)
- Fix lint: remove unused TableHeaderCellComponentProps import from XDSBaseTable
